Feat/vendor 업체담당자 type 변경 (#6)
* [Vendor]Fix: 업체담당 userId long type변경
* [Hub]Fix: 업체담당 userId long type변경

+++ hub/src/main/java/com/logistics/hub/application/dto/HubReadResDto.java
@@ -22,7 +22,7 @@ public class HubReadResDto {
 	private UUID id;
 	private String name;
 	private String address;
-	private int managerId;
+	private long managerId;
 
 	@JsonSerialize(using = LocalDateTimeSerializer.class)
 	@JsonDeserialize(using = LocalDateTimeDeserializer.class)
